added logo and fixed bugs

+++ ELCare/app/src/main/java/com/example/elcare/fragments/ChatFragment.java
@@ -196,8 +196,10 @@ public class ChatFragment extends Fragment {
                 Thread deleteSess = new Thread(new Runnable() {
                     @Override
                     public void run() {
-                        DeleteSessionOptions delOp = new DeleteSessionOptions.Builder(assID, sessionId).build();
-                        assService.deleteSession(delOp).execute();
+                        if(sessionId != null){
+                            DeleteSessionOptions delOp = new DeleteSessionOptions.Builder(assID, sessionId).build();
+                            assService.deleteSession(delOp).execute();
+                        }
                     }
                 });
                 deleteSess.start();
